Copy the location data to the new file
